+++ lib/domain/thumbnails.dart
@@ -16,15 +16,14 @@ Stream<List<Uint8List>> generateTrimThumbnails(
 
   for (int i = 1; i <= quantity; i++) {
     try {
-      final Uint8List? bytes = await VideoThumbnail.thumbnailData(
+      final bytes = await VideoThumbnail.thumbnailData(
         imageFormat: ImageFormat.JPEG,
         video: path,
         timeMs: (eachPart * i).toInt(),
         quality: quality,
       );
-      if (bytes != null) {
-        byteList.add(bytes);
-      }
+
+      byteList.add(bytes);
     } catch (e) {
       debugPrint(e.toString());
     }
@@ -74,7 +73,7 @@ Future<CoverData> generateSingleCoverThumbnail(
   int timeMs = 0,
   int quality = 10,
 }) async {
-  final Uint8List? thumbData = await VideoThumbnail.thumbnailData(
+  final thumbData = await VideoThumbnail.thumbnailData(
     imageFormat: ImageFormat.JPEG,
     video: filePath,
     timeMs: timeMs,