+++ lib/ui/trim/trim_slider_painter.dart
@@ -27,13 +27,10 @@ class TrimSliderPainter extends CustomPainter {
     // DRAW LEFT AND RIGHT BACKGROUNDS
     // extract [rect] trimmed area from the canvas
     canvas.drawPath(
-      Path.combine(
-        PathOperation.difference,
-        Path()..addRect(Rect.fromLTWH(0, 0, size.width, size.height)),
-        Path()
-          ..addRRect(rrect)
-          ..close(),
-      ),
+      Path()
+        ..fillType = PathFillType.evenOdd
+        ..addRect(Rect.fromLTWH(0, 0, size.width, size.height))
+        ..addRRect(rrect),
       background,
     );
 
@@ -91,30 +88,32 @@ class TrimSliderPainter extends CustomPainter {
     required Offset centerRight,
     required double halfLineWidth,
   }) {
+    // DRAW TOP AND BOTTOM LINES
     canvas.drawPath(
-      Path.combine(
-        PathOperation.union,
-        // DRAW TOP AND BOTTOM LINES
-        Path()
-          ..addRect(Rect.fromPoints(
+      Path()
+        ..addRect(
+          Rect.fromPoints(
             rect.topLeft,
             rect.topRight - Offset(0.0, style.lineWidth),
-          ))
-          ..addRect(
-            Rect.fromPoints(
-              rect.bottomRight + Offset(0.0, style.lineWidth),
-              rect.bottomLeft,
-            ),
           ),
-        // DRAW EDGES
-        getEdgesBarPath(
-          size,
-          centerLeft: centerLeft,
-          centerRight: centerRight,
-          halfLineWidth: halfLineWidth,
+        )
+        ..addRect(
+          Rect.fromPoints(
+            rect.bottomRight + Offset(0.0, style.lineWidth),
+            rect.bottomLeft,
+          ),
         ),
-      ),
       edges,
+    );
+
+    // DRAW EDGES
+    paintEdgesBarPath(
+      canvas,
+      size,
+      edges: edges,
+      centerLeft: centerLeft,
+      centerRight: centerRight,
+      halfLineWidth: halfLineWidth,
     );
 
     paintIcons(canvas, centerLeft: centerLeft, centerRight: centerRight);
@@ -122,102 +121,101 @@ class TrimSliderPainter extends CustomPainter {
     paintIndicator(canvas, size);
   }
 
-  Path getEdgesBarPath(
+  void paintEdgesBarPath(
+    Canvas canvas,
     Size size, {
+    required Paint edges,
     required Offset centerLeft,
     required Offset centerRight,
     required double halfLineWidth,
   }) {
     if (style.borderRadius == 0) {
-      return Path()
-        // LEFT EDGE
-        ..addRect(
-          Rect.fromCenter(
-            center: centerLeft,
-            width: style.edgesSize,
-            height: size.height + style.lineWidth * 2,
+      canvas.drawPath(
+        Path()
+          // LEFT EDGE
+          ..addRect(
+            Rect.fromCenter(
+              center: centerLeft,
+              width: style.edgesSize,
+              height: size.height + style.lineWidth * 2,
+            ),
+          )
+          // RIGTH EDGE
+          ..addRect(
+            Rect.fromCenter(
+              center: centerRight,
+              width: style.edgesSize,
+              height: size.height + style.lineWidth * 2,
+            ),
           ),
-        )
-        // RIGTH EDGE
-        ..addRect(
-          Rect.fromCenter(
-            center: centerRight,
-            width: style.edgesSize,
-            height: size.height + style.lineWidth * 2,
-          ),
-        );
+        edges,
+      );
     }
 
     final borderRadius = Radius.circular(style.borderRadius);
 
-    /// Return left and right edges, with a reversed border radius on the inside of the rect
-    return Path()
+    /// Left and right edges, with a reversed border radius on the inside of the rect
+    canvas.drawPath(
       // LEFT EDGE
-      ..addPath(
-        Path.combine(
-          PathOperation.difference,
-          Path()
-            ..addRRect(
-              RRect.fromRectAndCorners(
-                Rect.fromLTWH(
-                  centerLeft.dx - halfLineWidth,
-                  -style.lineWidth,
-                  style.edgeWidth + style.borderRadius,
-                  size.height + style.lineWidth * 2,
-                ),
-                topLeft: borderRadius,
-                bottomLeft: borderRadius,
-              ),
+      Path()
+        ..fillType = PathFillType.evenOdd
+        ..addRRect(
+          RRect.fromRectAndCorners(
+            Rect.fromLTWH(
+              centerLeft.dx - halfLineWidth,
+              -style.lineWidth,
+              style.edgeWidth + style.borderRadius,
+              size.height + style.lineWidth * 2,
             ),
-          Path()
-            ..addRRect(
-              RRect.fromRectAndCorners(
-                Rect.fromLTWH(
-                  centerLeft.dx + halfLineWidth,
-                  0.0,
-                  style.borderRadius,
-                  size.height,
-                ),
-                topLeft: borderRadius,
-                bottomLeft: borderRadius,
-              ),
+            topLeft: borderRadius,
+            bottomLeft: borderRadius,
+          ),
+        )
+        ..addRRect(
+          RRect.fromRectAndCorners(
+            Rect.fromLTWH(
+              centerLeft.dx + halfLineWidth,
+              0.0,
+              style.borderRadius,
+              size.height,
             ),
+            topLeft: borderRadius,
+            bottomLeft: borderRadius,
+          ),
         ),
-        Offset.zero,
-      )
+      edges,
+    );
+
+    canvas.drawPath(
       // RIGHT EDGE
-      ..addPath(
-        Path.combine(
-          PathOperation.difference,
-          Path()
-            ..addRRect(
-              RRect.fromRectAndCorners(
-                Rect.fromLTWH(
-                  centerRight.dx - halfLineWidth - style.borderRadius,
-                  -style.lineWidth,
-                  style.edgeWidth + style.borderRadius,
-                  size.height + style.lineWidth * 2,
-                ),
-                topRight: borderRadius,
-                bottomRight: borderRadius,
-              ),
+      Path()
+        ..fillType = PathFillType.evenOdd
+        ..addRRect(
+          RRect.fromRectAndCorners(
+            Rect.fromLTWH(
+              centerRight.dx - halfLineWidth - style.borderRadius,
+              -style.lineWidth,
+              style.edgeWidth + style.borderRadius,
+              size.height + style.lineWidth * 2,
             ),
-          Path()
-            ..addRRect(
-              RRect.fromRectAndCorners(
-                Rect.fromLTWH(
-                  centerRight.dx - halfLineWidth - style.borderRadius,
-                  0.0,
-                  style.borderRadius,
-                  size.height,
-                ),
-                topRight: borderRadius,
-                bottomRight: borderRadius,
-              ),
+            topRight: borderRadius,
+            bottomRight: borderRadius,
+          ),
+        )
+        ..addRRect(
+          RRect.fromRectAndCorners(
+            Rect.fromLTWH(
+              centerRight.dx - halfLineWidth - style.borderRadius,
+              0.0,
+              style.borderRadius,
+              size.height,
             ),
+            topRight: borderRadius,
+            bottomRight: borderRadius,
+          ),
         ),
-        Offset.zero,
-      );
+      edges,
+    );
   }
 
   void paintCircle(